+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 **1.0.1**
 * Fixed type errors for Dart 2
+* Fixed modal doesn't open when click element
 
 **1.0.0**
 * Dart 2 required